important setup
next step:

use the login modal pattern to send the todo data

or watch for textbox changes and click an add button to send the todo data

and check the foundSession.user and attach a user id to the todo data that you send

and jk jk dont watch

v-model:new_todo_date and send off with the add button click

there is a store

store ... state ... - userToken - user
..... this.$store.state.userToken ....... axios.get(url, config) {headers: {token : userToken} ........

so yeah do that sort of pattern I guess

dont do foundSession???? do token instead? Idk I have a misunderstanding of what a session is and what a token is.